+++ src/index.js
@@ -15,6 +15,16 @@ import * as serviceWorker from "./serviceWorker";
 
 const store = configureStore();
 
+if(config.app.googleAnalyticsEnabled === 'true'){
+  const tagManagerArgs = {
+    gtmId: config.app.googleAnalyticsContainerId
+  }
+
+  console.log("Activating Google Tag Manager");
+
+  TagManager.initialize(tagManagerArgs);
+}
+
 ReactDOM.render(
   <React.StrictMode>
     <Provider store={store}>
